refactor(Bio): rename
- change bio section to education section
- change splash section to bio section
