+++ frontend/src/components/sections/Bio.tsx
@@ -1,0 +1,6 @@
+import { FC } from 'react';
+import { Section } from '../../common/styles';
+
+export const BioSection: FC = () => {
+  return <Section id='bio'>Bio Section</Section>;
+};
