+++ frontend/src/components/sections/Education/CourseraCourseCard.tsx
