+++ frontend/src/components/sections/Education/UdemyCourseCard.tsx
