--- frontend/src/components/sections/Splash.tsx
@@ -1,6 +1,0 @@
-import { FC } from 'react';
-import { Section } from '../../common/styles';
-
-export const SplashSection: FC = () => {
-  return <Section id='splash'>Splash Section</Section>;
-};
